Change bg-color of nav and make font smaller
Make the background of navigation bar darker. Decrese font size of
header logo to 22px and navigations links to 16px. Also add a little
border below the navigation bar.

Fonts in navigation bar was looking a little bigger than in should be.
Change the background to experiment with the colorsheme.